# optimize int4 gemv kernel with cuda