--- a/onnxruntime/python/tools/kernel_explorer/kernels/matmul_4bits.py
+++ b/onnxruntime/python/tools/kernel_explorer/kernels/matmul_4bits.py
@@ -109,7 +109,7 @@ def profile():
     dims_m = [1]
     for dt in dtypes:
         for m in dims_m:
-            for n, k in ((4096, 4096), (4096, 12288), (12288, 4096)):
+            for n, k in ((4096, 4096), (4096, 12288), (12288, 4096), (4096, 11008), (11008, 4096),(2*11008, 4096),):
                 profile_with_args(m, n, k, dt, False)
                 print()