### Description
optimize gemv kernel:

1. unroll reduction to improve memory bandwidth
2. leverage 4bits to float16 tricks to save instrutions

| m | n     | k     | symmetric | latency before(us) | latency after(us) |
| - | ----- | ----- | --------- | ------------------ | ----------------- |
| 1 | 4096  | 4096  | TRUE      | 15.54              | 8.82              |
| 1 | 4096  | 4096  | FALSE     | 15.84              | 9.89              |
| 1 | 4096  | 11008 | TRUE      | 42.44              | 19.4              |
| 1 | 4096  | 11008 | FALSE     | 44.42              | 21.48             |
| 1 | 11008 | 4096  | TRUE      | 34.65              | 17.46             |
| 1 | 11008 | 4096  | FALSE     | 35.76              | 20.87             |
| 1 | 12288 | 4096  | TRUE      | 39.27              | 19.73             |
| 1 | 12288 | 4096  | FALSE     | 40.91              | 25.2              |
| 1 | 22016 | 4096  | TRUE      | 65.78              | 38.81             |
| 1 | 22016 | 4096  | FALSE     | 67.98              | 48.36             |